requirements.txt
future==0.18.2
python-dateutil==2.8.1
requests==2.25.1
pandas==1.3.4
googletrans==4.0.0rc1
flask==2.0.2